# kv: reenable the local call optimization by default