@petermattis 2016-04-13
I'm planning on investigating the discrepancy in the `Scan1` benchmarks next:

```
KVScan1_Native-32         324µs ± 3%    61µs ± 2%  -81.18%  (p=0.000 n=10+10)
KVScan1_SQL-32            535µs ± 2%   198µs ± 1%  -63.02%   (p=0.000 n=10+9)
```