@tschottdorf 2016-04-13
LGTM. Add benchmark results to commit message?