@petermattis 2016-04-13
I usually leave the benchmarks out of the commit messages. There are so many that are affected by this change. Do I show them all? Do I highlight the most affected? The least affected?