@tschottdorf 2016-04-13
I like putting them in (the ones that you've shown are enough), but not a stickler about it. Feel free to leave them out - stands to reason that they don't belong to the description of the change and are mostly a long block that gets into the way of casually reading through the history.