@spencerkimball 2016-04-14
Nice speedups.